Moved mouse over related code to subclass 'MouseOverMouseManager' of class 'DefaultMouseManager' so that it can be used optionally.
